Current Progress
This excludes the commit with the comment for now. The current issues are: Sometimes, after creating a profile, it asks you to put in your first and last name again after you submit it. Also, the webpage sometimes doesnt refresh properly and it requires manual refresh to see the new photo.